Develop

+++ Groomers.Services/CustomerService.cs
@@ -1,6 +1,8 @@
 ﻿using Groomers.Data;
 using Groomers.Models;
 using Groonmers.Data;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNet.Identity.EntityFramework;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -13,12 +15,17 @@ namespace Groomers.Services
     {
         private readonly Guid _userID;
         private readonly ApplicationDbContext _context;
+        private ApplicationUser _appUser;
+
+        //create a constructor that passes in that user and assigns to the field
+        //in our controller, we'll use similar syntax to get the Guid, but were going to use the .Find method to return the Application User
 
         public CustomerService(Guid userID)
         {
             _context = new ApplicationDbContext();
             _userID = userID;
         }
+
 
         //CREATE    
 
@@ -75,7 +82,7 @@ namespace Groomers.Services
 
             var customerList =
                 entityList
-                 .Where(e => e.UserID == _userID)
+                 //.Where(e => e.UserID == _userID)
                  .Select(
                     e =>
                         new CustomerListItem
@@ -94,9 +101,9 @@ namespace Groomers.Services
 
             return (customerList);
         }
-      
 
-        public CustomerDetail GetCustomerByCurrentUserId()
+
+        public CustomerListItem GetCustomerByCurrentUserId()
         {
             var entity = _context.Customers.Single(e => e.UserID == _userID);
 
@@ -110,21 +117,17 @@ namespace Groomers.Services
                 petList.Add(new PetListItem
                 {
                     PetID = pet.PetID,
-                    Name = pet.Name
+                    Name = $"{pet.Name}, "
                 });
             }
 
-            var model = new CustomerDetail
+            var model = new CustomerListItem
             {
                 PersonID = entity.PersonID,
-                FirstName = entity.FirstName,
-                LastName = entity.LastName,
-                StreetAddress = entity.StreetAddress,
-                City = entity.City,
-                State = entity.State,
-                ZipCode = entity.ZipCode,
+                FullName = entity.FullName,
                 PhoneNumber = entity.PhoneNumber,
                 Email = entity.Email,
+                ProfileCreationDate = entity.ProfileCreationDate,
                 Pets = petList,
             };
 
@@ -166,6 +169,7 @@ namespace Groomers.Services
             }
 
             List<AppointmentListItem> appList = new List<AppointmentListItem>();
+            List<AppointmentListItem> appListSorted = new List<AppointmentListItem>();
             foreach (var app in entity.Appointments)
             {
                 appList.Add(new AppointmentListItem
@@ -173,8 +177,11 @@ namespace Groomers.Services
                     AppointmentDate = app.AppointmentDate,
                     StartTime = app.StartTime,
                     PetID = app.PetID,
-                }); 
+                    PetName = app.Pet.Name,
+                });
+
             }
+            appListSorted = appList.OrderByDescending(x => x.AppointmentDate).ToList();
 
             var model = new CustomerDetail
             {
@@ -188,7 +195,7 @@ namespace Groomers.Services
                 PhoneNumber = entity.PhoneNumber,
                 Email = entity.Email,
                 Pets = petList,
-                Appointments = appList,
+                Appointments = appListSorted,
             };
             return (model);
 
@@ -218,6 +225,30 @@ namespace Groomers.Services
             }
         }
 
+        public bool AdminUpdateCustomer(CustomerEdit model)
+        {
+            using (_context)
+            {
+                var entity =
+                      _context
+                      .Customers
+                      .Single(e => e.PersonID == model.PersonID);
+
+                entity.PersonID = model.PersonID;
+                entity.FirstName = model.FirstName;
+                entity.LastName = model.LastName;
+                entity.StreetAddress = model.StreetAddress;
+                entity.City = model.City;
+                entity.State = model.State;
+                entity.ZipCode = model.ZipCode;
+                entity.PhoneNumber = model.PhoneNumber;
+                entity.Email = model.Email;
+                entity.ProfileModifiedDate = DateTimeOffset.Now;
+
+                return _context.SaveChanges() == 1;
+            }
+        }
+
         public bool DeleteCustomer(int id)
         {
             using (_context)
@@ -226,11 +257,71 @@ namespace Groomers.Services
                     _context
                     .Customers
                     .Single(e => e.PersonID == id && e.UserID == _userID);
-                _context.Customers.Remove(entity);
+                //_context.Customers.Remove(entity);
+
+                foreach (var app in entity.Appointments)
+                {
+                    app.PetID = null;
+                    app.PersonID = null;
+                    app.IsAvailable = true;
+
+                }
+                _context.SaveChanges();
+
+                if (entity.Appointments.Count != 0)
+                {
+                    return false;
+                }
+                else
+                    _context.Customers.Remove(entity);
+
+
+                _context.SaveChanges();
+
+                //check context to see if Pets contains an entity for the pet we deleted's Id
+                foreach (var customer in _context.Customers)
+                {
+                    if (customer.PersonID == id)
+                    {
+                        return false;
+                    }
+
+                }
+
+              //  DeleteAppUser();
+
 
                 return _context.SaveChanges() == 1;
+
             }
         }
+
+        public bool DeleteAppUser()
+        {
+            // var user = _context.Users.Find(_userID.ToString());
+            var userManager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(_context));
+
+            var roles = userManager.GetRoles(_userID.ToString());
+
+            //if role is admin, don't let user delete themself
+
+            foreach (var role in roles)
+            {
+
+                if (role != "Admin")
+                {
+                   // _appUser = role; 
+                    var result = userManager.Delete(_appUser);
+
+                    if (!result.Succeeded) return false;
+                    return true;
+                }
+                return false;
+            }
+            return false;
+
+        }
+        //after you run the initial delete method in the controller then call in the 
 
     }
 }
