+++ Groomers.Services/PetService.cs
@@ -12,14 +12,43 @@ namespace Groomers.Services
     public class PetService
     {
         private readonly ApplicationDbContext _context;
+
         private readonly Guid _userID;
+
+        // create a field to hold adminDog
+       // private readonly int _adminDog;
+
+        //create a method that get's adminDog
+        public PetListItem GetAdminDog(int id = 12)
+        {
+            using (_context)
+            {
+                var entity =
+                    _context
+                    .Pets
+                    .Single(e => e.PetID == id);
+
+                return
+                new PetListItem
+                {
+                    PetID = entity.PetID,
+                    Name = entity.Name,
+                };
+            }
+        }
+
+        //assigns the field it's value
+
+        // then you can just call that field.
+
 
         public PetService() { }
 
         public PetService(Guid userID)
         {
             _context = new ApplicationDbContext();
-            _userID = userID; 
+            _userID = userID;
+            // _adminDog = GetAdminDog(); 
         }
 
 
@@ -47,6 +76,26 @@ namespace Groomers.Services
             return _context.SaveChanges() == 1;
         }
 
+        public bool AdminCreatePet(PetCreate model)
+        {
+
+            var entity =
+                new Pet()
+                {
+                    Name = model.Name,
+                    SizeOfDog = model.SizeOfDog,
+                    IsHairLong = model.IsHairLong,
+                    SpecialRequest = model.SpecialRequest,
+                    Birthday = model.Birthday,
+                    DateAdded = DateTimeOffset.Now,
+                    PersonID = model.PersonID,
+                };
+
+            _context.Pets.Add(entity);
+
+            return _context.SaveChanges() == 1;
+        }
+
         public IEnumerable<PetListItem> GetPets()
         {
             var entityList = _context.Pets.ToList();
@@ -62,12 +111,12 @@ namespace Groomers.Services
                             SizeOfDog = e.SizeOfDog,
                             IsHairLong = e.IsHairLong,
                             SpecialRequest = e.SpecialRequest,
-                            DateAdded= e.DateAdded,
-                            PersonID = e.PersonID,
-                        }).ToList(); 
+                            DateAdded = e.DateAdded,
+                            OwnerName = e.Person.FullName,
+                        }).ToList();
 
             return (petList);
-            
+
         }
 
         public IEnumerable<PetListItem> GetPetsByUserID()
@@ -89,7 +138,7 @@ namespace Groomers.Services
                             IsHairLong = e.IsHairLong,
                             SpecialRequest = e.SpecialRequest,
                             DateAdded = e.DateAdded,
-                            PersonID = e.PersonID,
+                            //PersonID = e.PersonID,
                         }).ToList();
 
             return (petList);
@@ -104,7 +153,7 @@ namespace Groomers.Services
                     _context
                     .Pets
                     .Single(e => e.PetID == id);
-                
+
                 return
                 new PetDetail
                 {
@@ -113,8 +162,8 @@ namespace Groomers.Services
                     SizeOfDog = entity.SizeOfDog,
                     SpecialRequest = entity.SpecialRequest,
                     Birthday = entity.Birthday,
-                   // PersonID = entity.PersonID,
-                    FullName = entity.Person.FullName, 
+                    // PersonID = entity.PersonID,
+                    FullName = entity.Person.FullName,
                     DateAdded = entity.DateAdded,
                     DateModified = entity.DateModified,
                     Appointments = entity.Appointments.Select(app => new AppointmentDetails
@@ -142,26 +191,57 @@ namespace Groomers.Services
                 entity.SpecialRequest = model.SpecialRequest;
                 entity.DateModified = DateTimeOffset.Now;
 
-                return _context.SaveChanges() == 1; 
+                return _context.SaveChanges() == 1;
 
             }
         }
 
         public bool DeletePet(int id)
         {
+            // get currentusercustomerid
+            //make a variable of the customer's "adminDog"
             using (_context)
             {
                 var entity =
                     _context
                     .Pets
                     .Single(e => e.PetID == id);
-                    _context.Pets.Remove(entity);
 
-                return _context.SaveChanges() == 1; 
+                //foreach through entity's Appointments, change each Appointment's PetID to adminDog.PetID
+
+                foreach (var app in entity.Appointments)
+                {
+                    app.PetID = null;
+                    app.PersonID = null;
+                    app.IsAvailable = true;
+
+                }
+                _context.SaveChanges();
+
+                if (entity.Appointments.Count != 0)
+                {
+                    return false;
+                }
+                else
+                    _context.Pets.Remove(entity);
+                    
+                _context.SaveChanges();
+
+                //check context to see if Pets contains an entity for the pet we deleted's Id
+                foreach (var pet in _context.Pets)
+                {
+                    if (pet.PetID == id)
+                    {
+                        return false;
+                    }
+                }
+                return _context.SaveChanges() == 1;
+
+
             }
         }
 
-        
+
     }
 
 }
